fix: 20 critical training bug fixes (rounds 1-5)
NaN stability:
- Force LinOSSBlock to float32 (SSM values overflow fp16)
- Switch AMP to bfloat16 (range 3.39e38 vs fp16's 65504)
- Move OT loss outside AMP (Sinkhorn needs fp32)
- Increase OT regularization 0.05→1.0 for 128-dim embeddings
- Soft-clamp CFL ratio in DampedLayer for D-LinOSS stability
- Add NaN guard in trainer (skip batch on NaN/Inf loss)
- MoE gating softmax in float32 for AMP safety

Masking:
- Add learnable mask_token, apply mask BEFORE encoder
  (was passing unmasked spectrum → model learned identity)

Training loop:
- Grad accumulation: only log/validate/save on actual optimizer steps
- Skip step 0 for all periodic actions (0%N==0 is always true)
- Average losses over accumulation window instead of single sub-step
- Account for grad_accum in throughput calculation

VIB & losses:
- Replace KL-to-uniform with GradientReversal + cross-entropy
- GradientReversal uses x.clone() for DataParallel safety
- Apply physics loss to reconstruction (was applied to ground truth)
- Increase VIB weight 0.05→0.15

DataParallel:
- Convert domain list[str] to integer tensor before model.forward()
  so DataParallel scatters correctly across GPUs
- Both embedding classes accept tensor domain input

Optimizer:
- Exclude norms, biases, embeddings from weight decay
- Add eta_min to cosine schedule (don't decay LR to zero)
- Increase grad_clip 1.0→5.0

+++ src/models/moe.py
@@ -187,7 +187,7 @@ class MixtureOfExperts(nn.Module):
 
         # Top-k selection
         top_k_logits, top_k_indices = logits.topk(self.top_k, dim=-1)
-        top_k_gates = F.softmax(top_k_logits, dim=-1)
+        top_k_gates = F.softmax(top_k_logits.float(), dim=-1).to(top_k_logits.dtype)
 
         # Create sparse gate tensor (match softmax output dtype for AMP compatibility)
         gates = torch.zeros_like(logits, dtype=top_k_gates.dtype)
